Colors v1 - setting up a light source
Created two files: colors1.cpp and colors2.cpp

colors1.cpp - uses two VAOs and shaders, one for the general scene objects and one for the light source to render two different cubes to the screen. The larger, colored cube is rendered from the first VAO/shader, while the smaller white cube is the light source rendered from the second VAO/shader. This allows dynamic changing of lighting in the main scene without changing properties of the light source that aren't meant to change with the scene.

Also made a minor change to the exercise2.cpp file from the last chapter (camera).
-Updated the header comment at top to what the exercise was about, since it was originally copied from camera5.cpp and wasn't updated