Light Casters v4 - Spotlight w/ Smoother Edges
Implemented a line of code I found from a commenter on LearnOpenGL (Andrei Despinoiu) to make the edges of the flashlight/spotlight smoother.

Also experimented with getting a "cookie" texture overlayed on the spotlight to simulate a real flashlight, but couldn't figure it out. Going to revisit this once I know more